+++ packages/styled-components/components/link/index.js
@@ -15,6 +15,7 @@ import * as defaultStyles from './themes/default';
 const Link = (props) => {
   const {
     children,
+    className,
     href,
     onClick,
     rel,
@@ -38,6 +39,7 @@ const Link = (props) => {
 
   return (
     <LinkWrapper
+      className={className}
       href={relativeUrl || href}
       onClick={onClick || defaultOnClick}
       rel={rel}
@@ -50,6 +52,7 @@ const Link = (props) => {
 };
 
 Link.defaultProps = {
+  className: '',
   onClick: false,
   rel: '',
   style: {},
@@ -62,6 +65,10 @@ Link.propTypes = {
    * Child nodes
    */
   children: PropTypes.node.isRequired,
+  /**
+   * Class name.
+   */
+  className: PropTypes.string,
   /**
    * Destination for anchor tag (`href` attribute)
    */
@@ -81,7 +88,10 @@ Link.propTypes = {
   /**
    * CSS styles.
    */
-  style: PropTypes.object,
+  style: PropTypes.oneOfType([
+    PropTypes.array,
+    PropTypes.object,
+  ]),
   /**
    * Anchor target.
    */
@@ -95,5 +105,7 @@ Link.propTypes = {
 const themeMap = {
   default: defaultStyles,
 };
+
+export { Link as PureLink };
 
 export default withThemes(themeMap)(Link);