+++ packages/styled-components/components/text/themes/default.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
 
 /* eslint-disable import/prefer-default-export */
-export const HTMLWrapper = styled.div``;
+export const TextWrapper = styled.div``;
 /* eslint-enable */