+++ packages/styled-components/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.18.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v2.18.0...v2.18.1) (2020-06-26)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [2.18.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v2.17.0...v2.18.0) (2020-06-26)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [2.17.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v2.16.0...v2.17.0) (2020-06-24)
+
+
+### Bug Fixes
+
+* **proptypes for styles:** update style to allow either an array or object prop type ([70575d4](https://github.com/alleyinteractive/irving/packages/styled-components/commit/70575d4bfed80f4785268c0a0d298f37890d7ba1))
+* **styled-components:** remove refs that error out with local deps in a new theme ([8d564cc](https://github.com/alleyinteractive/irving/packages/styled-components/commit/8d564cc1ab70eeedb15fe1a29413aff7d9a385ed))
+
+
+### Features
+
+* **adds a css grid solution.:** adds a sane default/option for developers wanting to use CSS Grid ([d092689](https://github.com/alleyinteractive/irving/packages/styled-components/commit/d092689c9c45d533dce1d28535145df6e024ce49))
+* **container component.:** if a <container> is empty, do not render it ([c5e72b9](https://github.com/alleyinteractive/irving/packages/styled-components/commit/c5e72b905d43301820f8e33d7abffb513ac5743a))
+* **image component.:** implement the first alpha of an Image component ([b13341e](https://github.com/alleyinteractive/irving/packages/styled-components/commit/b13341e62534812d009e390f4f3f244d040a80ed))
+* **multiple components:** update and tweak the functionality on some components ([6b20b4d](https://github.com/alleyinteractive/irving/packages/styled-components/commit/6b20b4d01ce8c8226bb3a8dd538c4d0e86c60fa1))
+
+
+
+
+
+# [2.16.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v2.15.0...v2.16.0) (2020-06-19)
+
+
+### Features
+
+* **byline component:** cleanup some of the Byline component logic ([a3fce14](https://github.com/alleyinteractive/irving/packages/styled-components/commit/a3fce14a0f237a06302d8010fdced1fcc442d9a1))
+* scaffold storybook stories for all components ([ec3e43d](https://github.com/alleyinteractive/irving/packages/styled-components/commit/ec3e43d32257b7c88827a5699cfbcba574e66739))
+* **update storybook config to build correctly and add <link /> story:** fixes build proccess ([771be6b](https://github.com/alleyinteractive/irving/packages/styled-components/commit/771be6b8004a83c5de854be14ad5b12bb53a08e5))
+
+
+
+
+
+# [2.15.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v2.14.0...v2.15.0) (2020-06-16)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
 # [2.14.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v2.14.0-beta.1...v2.14.0) (2020-06-15)
 
 **Note:** Version bump only for package @irvingjs/styled-components